fixed physics issues with t-rex
fixed physics issues with t-rex. all game objects with a joint
needed to be set in the dinobehaviour script so that when the
t-rex is reset (onepisodebegin) when all bodyparts are reset,
it's affects all the game object with joints (even those not
directly controlled). It is important to do this, otherwise, the
reset of only a subset of the body part (or game object with
joints) does not cause the others to be reset. hence the body
parts not reset was still expereinced rotation/motion at the
start of each episode.

also added a couple of variables int he dino behaviour script
to keep track of episode rewards, number of episodes, and
number of steps in each episode.

 Changes to be committed:
	modified:   Assets/DinoBehaviourScript.cs
	modified:   Assets/ML-Agents/Examples/FirstTRexScene/DinoScene.unity